+++ lib/features/groups/group_detail_screen.dart
@@ -16,7 +16,7 @@ import '../../shared/widgets/balance_hero_card.dart';
 import '../../shared/widgets/balance_share_card.dart';
 import '../../shared/widgets/breakdown_pie_chart.dart';
 import '../../shared/widgets/member_expense_breakdown_sheet.dart';
-import '../../shared/widgets/pairwise_debt_tile.dart';
+import '../../shared/widgets/open_debt_tile.dart';
 import '../../shared/widgets/section_header.dart';
 import '../../shared/widgets/user_avatar.dart';
 import '../expenses/add_expense_screen.dart';
@@ -315,7 +315,7 @@ class GroupDetailScreen extends ConsumerWidget {
                       : Column(
                           children: [
                             for (var i = 0; i < data.debts.length; i++) ...[
-                              PairwiseDebtTile(
+                              OpenDebtTile(
                                 debt: data.debts[i],
                                 users: users,
                                 currencyCode: data.group.currencyCode,